tools/digdag/requirements.txt
openshift==0.9.0
kubernetes==9.0.0
python-string-utils
ruamel.yaml
six
